start events api/model

+++ config/db/seed/user.sql
@@ -3,8 +3,8 @@ CREATE TABLE users (
     email VARCHAR(50),
     password VARCHAR(60),
     name VARCHAR(50),
-    facebook_id INT,
-    is_active INT DEFAULT 1,
+    facebook_id BIGINT,
+    is_active SMALLINT DEFAULT 1,
     created_at TIMESTAMP DEFAULT (NOW() AT TIME ZONE 'PST'),
     updated_at TIMESTAMP DEFAULT (NOW() AT TIME ZONE 'PST')
 );
